# src/main/resources/docs/dgbank-api-docs.yaml
openapi: 3.0.3
info:
  title: DG Bank API
  description: |
    API completa per il sistema bancario DG Bank.
    
    ## Autenticazione
    L'API utilizza JWT Bearer token per l'autenticazione. Dopo il login o la registrazione, includi il token nell'header Authorization:
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Token JWT: 24 ore di validità
    - Refresh token: 7 giorni di validità
    
    ## Formati Monetari
    Tutti gli importi sono gestiti con precisione decimale (15,2) per evitare errori di arrotondamento.
  version: 1.0.0
  contact:
    name: DG Bank Development Team
    email: dev@dgbank.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.dgbank.com
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Registrazione nuovo utente
      description: |
        Registra un nuovo utente nel sistema. L'utente deve essere maggiorenne (18+).
        La password viene crittografata con BCrypt prima del salvataggio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: "mario.rossi@email.com"
              password: "password123"
              firstName: "Mario"
              lastName: "Rossi"
              dateOfBirth: "1990-01-15"
              phoneNumber: "+39 333 1234567"
              documentType: "CARTA_IDENTITA"
              documentNumber: "AB1234567"
              documentExpiry: "2030-01-15"
              documentIssuer: "Comune di Milano"
              street: "Via Roma 123"
              city: "Milano"
              postalCode: "20100"
              province: "MI"
              country: "Italia"
      responses:
        '201':
          description: Utente registrato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Errore nei dati di registrazione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                email_exists:
                  summary: Email già registrata
                  value:
                    success: false
                    message: "Email già registrata: mario.rossi@email.com"
                document_exists:
                  summary: Documento già registrato
                  value:
                    success: false
                    message: "Documento già registrato"
                underage:
                  summary: Utente minorenne
                  value:
                    success: false
                    message: "L'utente deve essere maggiorenne"
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login utente
      description: Autentica un utente esistente e restituisce i token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "mario.rossi@email.com"
              password: "password123"
      responses:
        '200':
          description: Login effettuato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenziali non valide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Credenziali non valide"
  /auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Rinnovo token
      description: Rinnova il token di accesso utilizzando il refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token rinnovato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Refresh token non valido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/validate-token:
    post:
      tags:
        - Authentication
      summary: Validazione token
      description: Verifica la validità di un token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateTokenRequest'
      responses:
        '200':
          description: Risultato validazione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateTokenResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT ottenuto tramite login o registrazione.
        Durata: 24 ore per access token, 7 giorni per refresh token.
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - dateOfBirth
        - phoneNumber
        - documentType
        - documentNumber
        - documentExpiry
        - documentIssuer
        - street
        - city
        - postalCode
        - province
        - country
      properties:
        email:
          type: string
          format: email
          maxLength: 100
          description: Indirizzo email univoco
        password:
          type: string
          minLength: 8
          description: Password che verrà crittografata con BCrypt
        firstName:
          type: string
          maxLength: 50
          description: Nome dell'utente
        lastName:
          type: string
          maxLength: 50
          description: Cognome dell'utente
        dateOfBirth:
          type: string
          format: date
          description: Data di nascita (deve essere maggiorenne)
        phoneNumber:
          type: string
          maxLength: 20
          description: Numero di telefono
        documentType:
          $ref: '#/components/schemas/DocumentType'
        documentNumber:
          type: string
          maxLength: 50
          description: Numero del documento di identità (deve essere univoco)
        documentExpiry:
          type: string
          format: date
          description: Data di scadenza del documento
        documentIssuer:
          type: string
          maxLength: 100
          description: Ente che ha rilasciato il documento
        street:
          type: string
          description: Via e numero civico
        city:
          type: string
          maxLength: 100
          description: Città di residenza
        postalCode:
          type: string
          maxLength: 10
          description: Codice postale/CAP
        province:
          type: string
          maxLength: 50
          description: Provincia di residenza
        country:
          type: string
          maxLength: 50
          description: Paese di residenza
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Indirizzo email dell'utente
        password:
          type: string
          description: Password dell'utente
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Token di refresh ottenuto dal login
    ValidateTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Token JWT da validare
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Login successful"
        token:
          type: string
          description: Token JWT di accesso (24h)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: Token di refresh (7 giorni)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
    TokenResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          description: Nuovo token JWT di accesso
        refreshToken:
          type: string
          description: Token di refresh (invariato)
    ValidateTokenResponse:
      type: object
      properties:
        valid:
          type: boolean
          description: true se il token è valido
        email:
          type: string
          format: email
          description: Email estratta dal token (se valido)
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Messaggio di errore"

    # Entity models
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID univoco dell'utente
        email:
          type: string
          format: email
          description: Indirizzo email
        firstName:
          type: string
          description: Nome
        lastName:
          type: string
          description: Cognome
        fullName:
          type: string
          description: Nome completo (nome + cognome)
          example: "Mario Rossi"
        dateOfBirth:
          type: string
          format: date
          description: Data di nascita
        fullAddress:
          type: string
          description: Indirizzo completo formattato
          example: "Via Roma 123, 20100 Milano (MI), Italia"
        registeredDocument:
          type: string
          description: Tipo di documento registrato
          example: "Carta d'Identità"
...
    DocumentType:
      type: string
      enum:
        - CARTA_IDENTITA
        - PASSAPORTO
        - PATENTE
        - PERMESSO_SOGGIORNO
      description: |
        Tipologie di documento di identità supportate:
        * `CARTA_IDENTITA` - Carta d'Identità
        * `PASSAPORTO` - Passaporto
        * `PATENTE` - Patente di Guida
        * `PERMESSO_SOGGIORNO` - Permesso di Soggiorno
...
security:
  - bearerAuth: []

# Esempi di utilizzo
...